feat(security):  fix security 404
